created ScoreBoard container with its RowScore and RowEmpty components
